requirements.txt
Pillow
smbus2>=0.3.0

